requirements.txt
Flask==2.2.2
torch==2.0.0
torchvision==0.15.0
Pillow==9.4.0
gunicorn==20.1.0